requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
azure-storage-file-share==12.7.0
fastjsonschema==2.15.3
parsedatetime==2.6
